+++ src/ConfigBuilder.php
@@ -74,6 +74,9 @@ final class ConfigBuilder
         return $this;
     }
 
+    /**
+     * @deprecated Use `ApiServiceFactory` or `ApiServiceFactoryBuilder::setSoapOptions()` instead.
+     */
     public function setSoapOptions(SoapOptions $options): self
     {
         $this->options['soap_options'] = $options;